+++ README.md
@@ -150,11 +150,16 @@ All Fields except `<Info />` accept an `attr` property to bind them to a key in 
 * `<ColorPicker />`: renders as a popup color-picker, useful for CSS color hex value strings
 * `<Flaglist />`: renders as a list of checkboxes, useful for `+`-joined flag lists like `data[].mode`
 * `<MultiFormatTextEditor />`: renders as a WYSIWYG editor, useful for text like `layout.title`
-* `<Button/>`: simple button component, useful when combined with `<SingleSidebarItem/>` to add as menu item
 
 <p align="center">
   <img src="examples/components.png" alt="Components" width="432" height="692" border="1">
 </p>
+
+### Widgets
+
+Simple component that takes in props and renders.
+
+* `<Button/>`: simple button component, useful when combined with `<SingleSidebarItem/>` to add as menu item
 
 ### Special-Purpose Containers
