+++ src/components/containers/index.js
@@ -7,7 +7,7 @@ import TraceAccordion from './TraceAccordion';
 import TraceMarkerSection from './TraceMarkerSection';
 import {LayoutPanel, AxesFold} from './derived';
 import TraceRequiredPanel from './TraceRequiredPanel';
-import SingleSidebarItem from '../sidebar/SingleSidebarItem';
+import SingleSidebarItem from './SingleSidebarItem';
 
 export {
   AnnotationAccordion,
