+++ src/components/widgets/index.js
@@ -1,0 +1,2 @@
+import Button from './Button';
+export {Button};
